+++ src/argilla/client/feedback/dataset.py
@@ -12,9 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import json
 import logging
-import tempfile
 from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Literal, Optional, Union
 from uuid import UUID
 
@@ -25,12 +23,11 @@ from pydantic import (
 from tqdm import tqdm
 
 import argilla as rg
-from argilla.client.feedback.config import FeedbackDatasetConfig
 from argilla.client.feedback.constants import (
     FETCHING_BATCH_SIZE,
-    FIELD_TYPE_TO_PYTHON_TYPE,
     PUSHING_BATCH_SIZE,
 )
+from argilla.client.feedback.integrations.huggingface import HuggingFaceDatasetMixIn
 from argilla.client.feedback.schemas import (
     FeedbackRecord,
     FieldSchema,
@@ -72,7 +69,7 @@ if TYPE_CHECKING:
 _LOGGER = logging.getLogger(__name__)
 
 
-class FeedbackDataset:
+class FeedbackDataset(HuggingFaceDatasetMixIn):
     """Class to work with `FeedbackDataset`s either locally, or remotely (Argilla or HuggingFace Hub).
 
     Args:
@@ -760,248 +757,18 @@ class FeedbackDataset:
             >>> huggingface_dataset = dataset.format_as("datasets")
         """
         if format == "datasets":
-            from datasets import Dataset, Features, Sequence, Value
-
-            dataset = {"metadata": []}
-            features = {}
-            for field in self.fields:
-                if field.settings["type"] not in FIELD_TYPE_TO_PYTHON_TYPE.keys():
-                    raise ValueError(
-                        f"Field {field.name} has an unsupported type: {field.settings['type']}, for the moment"
-                        f" only the following types are supported: {list(FIELD_TYPE_TO_PYTHON_TYPE.keys())}"
-                    )
-                features[field.name] = Value(dtype="string", id="field")
-                if field.name not in dataset:
-                    dataset[field.name] = []
-            for question in self.questions:
-                if question.settings["type"] in ["text", "label_selection"]:
-                    value = Value(dtype="string")
-                elif question.settings["type"] == "rating":
-                    value = Value(dtype="int32")
-                elif question.settings["type"] == "ranking":
-                    value = Sequence({"rank": Value(dtype="uint8"), "value": Value(dtype="string")})
-                elif question.settings["type"] in "multi_label_selection":
-                    value = Sequence(Value(dtype="string"))
-                else:
-                    raise ValueError(
-                        f"Question {question.name} has an unsupported type: {question.settings['type']}, for the"
-                        " moment only the following types are supported: 'text', 'rating', 'label_selection',"
-                        " 'multi_label_selection', and 'ranking'."
-                    )
-                # TODO(alvarobartt): if we constraint ranges from 0 to N, then we can use `ClassLabel` for ratings
-                features[question.name] = Sequence(
-                    {
-                        "user_id": Value(dtype="string"),
-                        "value": value,
-                        "status": Value(dtype="string"),
-                    },
-                    id="question",
-                )
-                if question.name not in dataset:
-                    dataset[question.name] = []
-            features["external_id"] = Value(dtype="string", id="external_id")
-            dataset["external_id"] = []
-
-            for record in self.records:
-                for field in self.fields:
-                    dataset[field.name].append(record.fields[field.name])
-                for question in self.questions:
-                    if not record.responses:
-                        dataset[question.name].append(None)
-                        continue
-                    responses = []
-                    for response in record.responses:
-                        if question.name not in response.values:
-                            responses.append(None)
-                            continue
-                        if question.settings["type"] == "ranking":
-                            responses.append([r.dict() for r in response.values[question.name].value])
-                        else:
-                            responses.append(response.values[question.name].value)
-                    dataset[question.name].append(
-                        {
-                            "user_id": [r.user_id for r in record.responses],
-                            "value": responses,
-                            "status": [r.status for r in record.responses],
-                        }
-                    )
-                dataset["metadata"].append(json.dumps(record.metadata) if record.metadata else None)
-                dataset["external_id"].append(record.external_id or None)
-
-            if len(dataset["metadata"]) > 0:
-                features["metadata"] = Value(dtype="string")
-            else:
-                del dataset["metadata"]
-
-            return Dataset.from_dict(
-                dataset,
-                features=Features(features),
-            )
+            return self._huggingface_format(self)
         raise ValueError(f"Unsupported format '{format}'.")
 
     @requires_version("huggingface_hub")
     def push_to_huggingface(self, repo_id: str, generate_card: Optional[bool] = True, *args, **kwargs) -> None:
-        """Pushes the `FeedbackDataset` to the HuggingFace Hub. If the dataset has been previously pushed to the
-        HuggingFace Hub, it will be updated instead. Note that some params as `private` have no effect at all
-        when a dataset is previously uploaded to the HuggingFace Hub.
-
-        Args:
-            repo_id: the ID of the HuggingFace Hub repo to push the `FeedbackDataset` to.
-            generate_card: whether to generate a dataset card for the `FeedbackDataset` in the HuggingFace Hub. Defaults
-                to `True`.
-            *args: the args to pass to `datasets.Dataset.push_to_hub`.
-            **kwargs: the kwargs to pass to `datasets.Dataset.push_to_hub`.
-        """
-        import huggingface_hub
-        from huggingface_hub import DatasetCardData, HfApi
-        from packaging.version import parse as parse_version
-
-        if parse_version(huggingface_hub.__version__) < parse_version("0.14.0"):
-            _LOGGER.warning(
-                "Recommended `huggingface_hub` version is 0.14.0 or higher, and you have"
-                f" {huggingface_hub.__version__}, so in case you have any issue when pushing the dataset to the"
-                " HuggingFace Hub upgrade it as `pip install huggingface_hub --upgrade`."
-            )
-
-        if len(self) < 1:
-            raise ValueError(
-                "Cannot push an empty `rg.FeedbackDataset` to the HuggingFace Hub, please make sure to add at"
-                " least one record, via the method `add_records`."
-            )
-
-        hfds = self.format_as("datasets")
-        hfds.push_to_hub(repo_id, *args, **kwargs)
-
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".cfg", delete=False) as f:
-            f.write(
-                FeedbackDatasetConfig(fields=self.fields, questions=self.questions, guidelines=self.guidelines).json()
-            )
-            f.flush()
-
-            HfApi().upload_file(
-                path_or_fileobj=f.name,
-                path_in_repo="argilla.cfg",
-                repo_id=repo_id,
-                repo_type="dataset",
-                token=kwargs.get("token"),
-            )
-
-        if generate_card:
-            from argilla.client.feedback.card import (
-                ArgillaDatasetCard,
-                size_categories_parser,
-            )
-
-            card = ArgillaDatasetCard.from_template(
-                card_data=DatasetCardData(
-                    size_categories=size_categories_parser(len(self.records)),
-                    tags=["rlfh", "argilla", "human-feedback"],
-                ),
-                repo_id=repo_id,
-                argilla_fields=self.fields,
-                argilla_questions=self.questions,
-                argilla_guidelines=self.guidelines,
-                argilla_record=json.loads(self.records[0].json()),
-                huggingface_record=hfds[0],
-            )
-            card.push_to_hub(repo_id, repo_type="dataset", token=kwargs.get("token"))
+        return self._push_to_huggingface(self, repo_id, generate_card=generate_card, *args, **kwargs)
 
     @classmethod
     @requires_version("datasets")
     @requires_version("huggingface_hub")
     def from_huggingface(cls, repo_id: str, *args: Any, **kwargs: Any) -> "FeedbackDataset":
-        """Loads a `FeedbackDataset` from the HuggingFace Hub.
-
-        Args:
-            repo_id: the ID of the HuggingFace Hub repo to load the `FeedbackDataset` from.
-            *args: the args to pass to `datasets.Dataset.load_from_hub`.
-            **kwargs: the kwargs to pass to `datasets.Dataset.load_from_hub`.
-
-        Returns:
-            A `FeedbackDataset` loaded from the HuggingFace Hub.
-        """
-        import huggingface_hub
-        from datasets import DatasetDict, load_dataset
-        from huggingface_hub import hf_hub_download
-        from packaging.version import parse as parse_version
-
-        if parse_version(huggingface_hub.__version__) < parse_version("0.14.0"):
-            _LOGGER.warning(
-                "Recommended `huggingface_hub` version is 0.14.0 or higher, and you have"
-                f" {huggingface_hub.__version__}, so in case you have any issue when pushing the dataset to the"
-                " HuggingFace Hub upgrade it as `pip install huggingface_hub --upgrade`."
-            )
-
-        if "token" in kwargs:
-            auth = kwargs.pop("token")
-        elif "use_auth_token" in kwargs:
-            auth = kwargs.pop("use_auth_token")
-        else:
-            auth = None
-
-        hub_auth = (
-            {"use_auth_token": auth}
-            if parse_version(huggingface_hub.__version__) < parse_version("0.11.0")
-            else {"token": auth}
-        )
-        config_path = hf_hub_download(
-            repo_id=repo_id,
-            filename="argilla.cfg",
-            repo_type="dataset",
-            **hub_auth,
-        )
-        with open(config_path, "r") as f:
-            config = FeedbackDatasetConfig.parse_raw(f.read())
-
-        cls = cls(
-            fields=config.fields,
-            questions=config.questions,
-            guidelines=config.guidelines,
-        )
-
-        hfds = load_dataset(repo_id, use_auth_token=auth, *args, **kwargs)
-        if isinstance(hfds, DatasetDict) and "split" not in kwargs:
-            if len(hfds.keys()) > 1:
-                raise ValueError(
-                    "Only one dataset can be loaded at a time, use `split` to select a split, available splits"
-                    f" are: {', '.join(hfds.keys())}."
-                )
-            hfds = hfds[list(hfds.keys())[0]]
-
-        for index in range(len(hfds)):
-            responses = {}
-            for question in cls.questions:
-                if hfds[index][question.name] is None or len(hfds[index][question.name]) < 1:
-                    continue
-                for user_id, value, status in zip(
-                    hfds[index][question.name]["user_id"],
-                    hfds[index][question.name]["value"],
-                    hfds[index][question.name]["status"],
-                ):
-                    if user_id not in responses:
-                        responses[user_id] = {
-                            "user_id": user_id,
-                            "status": status,
-                            "values": {},
-                        }
-                    if question.settings["type"] == "ranking":
-                        value = [{"rank": r, "value": v} for r, v in zip(value["rank"], value["value"])]
-                    responses[user_id]["values"].update({question.name: {"value": value}})
-
-            metadata = None
-            if "metadata" in hfds[index] and hfds[index]["metadata"] is not None:
-                metadata = json.loads(hfds[index]["metadata"])
-
-            cls.__records.append(
-                FeedbackRecord(
-                    fields={field.name: hfds[index][field.name] for field in cls.fields},
-                    metadata=metadata,
-                    responses=list(responses.values()) or None,
-                    external_id=hfds[index]["external_id"],
-                )
-            )
-        del hfds
-        return cls
+        return cls._from_huggingface(repo_id, *args, **kwargs)
 
     def unify_responses(
         self,
