+++ src/argilla/client/feedback/integrations/__init__.py
@@ -11,8 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from argilla.client.feedback.card._dataset_card import ArgillaDatasetCard
-from argilla.client.feedback.card._parser import size_categories_parser
-
-__all__ = ["ArgillaDatasetCard", "size_categories_parser"]
